fix | socket to pipe error fixed | ClientSocket.cpp

+++ include/Request.hpp
@@ -44,7 +44,7 @@ public:
   size_t generalURI(std::string location) const;
   std::string combinePATH(Location target, size_t rate) const;
   void convertURI();
-  void eraseRequestBody(size_t start, size_t end);
+  void eraseRequestBody(int start, int end);
   void initRequest();
 
 private:
